helper functions to util
